linkplace-phase1-backend_20250915_102537/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Authentication and Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# OAuth2 and Social Login
authlib==1.2.1
httpx==0.25.2

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Background Tasks and Caching
celery==5.3.4
redis==5.0.1

# Environment and Configuration
python-dotenv==1.0.0

# Date and Time
python-dateutil==2.8.2

# Logging and Monitoring
structlog==23.2.0

# Testing utilities (for basic validation)
pytest==7.4.3
